Implement Penny Precision Hardening & Liquidation Dust Verification
- Modified `EconomicIndicatorTracker` to output strict integer (Penny) values for consumption, income, and average prices.
- Updated `EconomicIndicatorData` DTO to use `Optional[int]` for `avg_wage`, `avg_goods_price`, and `food_avg_price`.
- Verified `LiquidationManager` correctly handles pro-rata dust distribution (Zero-Sum).
- Updated integration tests to expect Pennies instead of Dollars.
- Added strict type assertions in `tests/test_economic_tracker_precision.py`.
- Ensured 100% test pass rate.

+++ simulation/dtos/api.py
@@ -88,10 +88,10 @@ class EconomicIndicatorData:
     run_id: int
     time: int
     unemployment_rate: Optional[float] = None
-    avg_wage: Optional[float] = None
-    food_avg_price: Optional[float] = None
+    avg_wage: Optional[int] = None # Changed for Reporting DTO Hardening (Pennies)
+    food_avg_price: Optional[int] = None # Changed for Reporting DTO Hardening (Pennies)
     food_trade_volume: Optional[float] = None
-    avg_goods_price: Optional[float] = None
+    avg_goods_price: Optional[int] = None # Changed for Reporting DTO Hardening (Pennies)
     total_production: Optional[float] = None
     total_consumption: Optional[int] = None
     total_household_assets: Optional[int] = None # Changed for Reporting DTO Hardening (Pennies)
